e-puck2 firmware cont'd
* add support for IR remote control:
- TIM11 used to read the RC5 protocol
- default speed used to move the robot with basic IR commands (1..9)
- receive commands @ 5 Hz
* IMU:
- use msgbus to publish accelerometer and gyro raw values @ 20 Hz
- in the main subscribe to the imu topic in order to print the sensor
values
* e-puck1.x wrapper library:
- corrected speed handling in order to move e-puck2 the same speed as
e-puck1.x
* corrected a bug in the speed setting from asercom

+++ src/epuck1x/Asercom.c
@@ -379,8 +379,8 @@ int run_asercom(void) {
                             while (e_getchar_uart2(&c2) == 0);
                         }
                         speedr = (unsigned char) c1 + ((unsigned int) c2 << 8);
-                        e_set_speed_left(speedl);
-                        e_set_speed_right(speedr);
+                        e_set_speed_left((int16_t)speedl);
+                        e_set_speed_right((int16_t)speedr);
                         break;
                     case 'E': // get motor speed
                         buffer[i++] = speedl & 0xff;
